TEST-service/requirements.txt
click
fastapi[all]
hydra-core
joblib
motor
omegaconf
pdf2image
Pillow
pymongo
PyMuPDF
python-multipart
urllib3
pandas
azure==4.0.0
azure-ai-formrecognizer==3.1.2
APScheduler
pytesseract
sqlalchemy
boto3
botocore
opentelemetry-instrumentation-fastapi
opentelemetry-distro
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
